requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
requests
beautifulsoup4
chromadb
sentence-transformers
Jinja2
python-multipart
